完成机构的增删改查功能; 暂时关闭登陆验证功能; 新增一些表设计

+++ src/main/java/tt/service/bussiness/impl/CompanyServiceImpl.java
@@ -1,11 +1,13 @@
 package tt.service.bussiness.impl;
 
+import com.alibaba.fastjson.JSONObject;
 import tt.dao.business.CompanyDaoI;
 import tt.model.business.Company;
 import tt.service.bussiness.CompanyServiceI;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -20,20 +22,39 @@ public class CompanyServiceImpl implements CompanyServiceI {
 
     @Override
     public Company get(int id) {
+        System.out.println("run get");
         return companyDao.get(Company.class, id);
     }
 
     @Override
-    public List<Company> list(Map<String, Object> params, int page, int PageSize) {
-        String hql = "from Company where typ in (:typ)";
-        List<Company> ret = companyDao.find(hql, params, page, PageSize);
+    public List<Company> list(Byte typ,String name, int page, int PageSize) {
+        StringBuilder hql = new StringBuilder("from Company WHERE 1=1");
+        Map<String,Object> params = new HashMap<>();
+        if(typ!=null&&typ!=0){
+            params.put("typ",typ);
+            hql.append(" AND typ=:typ ");
+        }
+        if(name!=null){
+            params.put("name","%"+name+"%");
+            hql.append(" AND name like :name ");
+        }
+        List<Company> ret = companyDao.find(hql.toString(), params, page, PageSize);
         return ret;
     }
 
     @Override
-    public long count(Map<String, Object> params) {
-        String sql = "select count(*) from Company where typ=:typ";
-        long ret = companyDao.count(sql, params);
+    public long count(Byte typ,String name) {
+        StringBuilder hql = new StringBuilder("select count(*) from Company WHERE 1=1");
+        Map<String,Object> params = new HashMap<>();
+        if(typ!=null&&typ!=0){
+            params.put("typ",typ);
+            hql.append(" AND typ=:typ ");
+        }
+        if(name!=null){
+            params.put("name","%"+name+"%");
+            hql.append(" AND name like :name ");
+        }
+        long ret = companyDao.count(hql.toString(), params);
         return ret;
     }
 
@@ -44,7 +65,10 @@ public class CompanyServiceImpl implements CompanyServiceI {
     }
 
     @Override
-    public int del(int[] ids) {
+    public int del(List<Integer> ids) {
+        if(ids==null||ids.size()<1){
+            return 0;
+        }
         Map<String, Object> params = new HashMap<String, Object>();
         params.put("ids", ids);
         return companyDao.executeHql("delete from Company where id in (:ids)", params);
